+++ .github/workflows/build.yml
@@ -1,12 +1,16 @@
-name: Build the library
+name: build
 
-on: [push]
+on: [push, pull_request]
 
 jobs:
   build:
     runs-on: ubuntu-latest
     steps:
     - uses: actions/checkout@v2
+
+    - uses: actions/setup-node@v1
+      with:
+        node-version: '10.x'
 
     - name: Install dependencies
       run: npm install
@@ -19,3 +23,8 @@ jobs:
 
     - name: Test
       run: npm run test
+
+    - name: Coveralls
+      uses: coverallsapp/github-action@master
+      with:
+        github-token: ${{ secrets.GITHUB_TOKEN }}
