+++ jest.config.js
@@ -1,5 +1,10 @@
 module.exports = {
   preset: 'ts-jest',
   testEnvironment: 'node',
-  rootDir: 'test'
+  collectCoverage: true,
+  collectCoverageFrom: ['src/**/*.ts'],
+  roots: [
+      'test',
+      'src'
+  ]
 };
